backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
Werkzeug==3.0.1
cryptography==41.0.7
python-dotenv==1.0.0

# AI 相关依赖
httpx>=0.25.0                    # HTTP客户端(用于调用LLM API)
qdrant-client>=1.7.0             # Qdrant向量数据库客户端
PyPDF2>=3.0.0                    # PDF解析
python-docx>=1.0.0               # Word文档解析
sentence-transformers>=2.2.0     # 本地embedding模型 (BGE)
torch>=2.0.0                     # PyTorch (sentence-transformers依赖)

# 可选依赖(按需安装)
# rank-bm25>=0.2.0               # BM25关键词检索（混合检索时使用）

